Added alphabetization by title to library functions
Added a basic function that sorts files by name into alphabetical
containers held by the MusicLibrary object, written at the skill level
of a high schooler who is really into python scripting. Please
refactor/end the life of this code as soon as possible; it's an
abomination. I just wanted to see if I could do it. Also ignore the fact
that the only music present is The Strokes. The files are named
conveniently.